refactor: split auth persistence helpers
